refactor(pkg): refactoring internal client package
The internalclient package now features improved organization with
clearer distinctions between files and interfaces. A full documentation
are also provided
