+++ pkg/api/internalclient/client.go
@@ -2,17 +2,32 @@ package internalclient
 
 import (
 	"errors"
-	"fmt"
 	"math"
 	"net"
 	"net/http"
-	"net/url"
 
 	resty "github.com/go-resty/resty/v2"
 	"github.com/hibiken/asynq"
-	"github.com/shellhub-io/shellhub/pkg/models"
 	"github.com/sirupsen/logrus"
 )
+
+type client struct {
+	http   *resty.Client
+	logger *logrus.Logger
+	asynq  *asynq.Client
+}
+
+type Client interface {
+	deviceAPI
+	namespaceAPI
+	billingAPI
+	sessionAPI
+	sshkeyAPI
+	firewallAPI
+}
+
+// Ensures the client implements Client.
+var _ Client = (*client)(nil)
 
 const (
 	DeviceUIDHeader = "X-Device-UID"
@@ -35,6 +50,7 @@ type Opt func(*Options) error
 
 func NewClient(opts ...Opt) Client {
 	httpClient := resty.New()
+	httpClient.SetBaseURL("http://api:8080")
 	httpClient.SetRetryCount(math.MaxInt32)
 	httpClient.AddRetryCondition(func(r *resty.Response, err error) bool {
 		if _, ok := err.(net.Error); ok { // if the error is a network error, retry.
@@ -44,12 +60,7 @@ func NewClient(opts ...Opt) Client {
 		return r.StatusCode() >= http.StatusInternalServerError && r.StatusCode() != http.StatusNotImplemented
 	})
 
-	c := &client{
-		host:   apiHost,
-		port:   apiPort,
-		scheme: apiScheme,
-		http:   httpClient,
-	}
+	c := &client{http: httpClient}
 
 	o := new(Options)
 	for _, opt := range opts {
@@ -92,74 +103,4 @@ func NewClientWithAsynq(uri string) Client {
 
 		return nil
 	})
-}
-
-type commonAPI interface {
-	ListDevices() ([]models.Device, error)
-	GetDevice(uid string) (*models.Device, error)
-	GetDeviceByPublicURLAddress(address string) (*models.Device, error)
-}
-
-type client struct {
-	scheme string
-	host   string
-	port   int
-	http   *resty.Client
-	logger *logrus.Logger
-	asynq  *asynq.Client
-}
-
-func (c *client) ListDevices() ([]models.Device, error) {
-	list := []models.Device{}
-	_, err := c.http.R().
-		SetResult(list).
-		Get(buildURL(c, "/api/devices"))
-
-	return list, err
-}
-
-func (c *client) GetDevice(uid string) (*models.Device, error) {
-	var device *models.Device
-	resp, err := c.http.R().
-		SetResult(&device).
-		Get(buildURL(c, fmt.Sprintf("/api/devices/%s", uid)))
-	if err != nil {
-		return nil, ErrConnectionFailed
-	}
-
-	switch resp.StatusCode() {
-	case 400:
-		return nil, ErrNotFound
-	case 200:
-		return device, nil
-	default:
-		return nil, ErrUnknown
-	}
-}
-
-func (c *client) GetDeviceByPublicURLAddress(address string) (*models.Device, error) {
-	httpClient := resty.New()
-
-	var device *models.Device
-	resp, err := httpClient.R().
-		SetResult(&device).
-		Get(buildURL(c, fmt.Sprintf("/internal/devices/public/%s", address)))
-	if err != nil {
-		return nil, ErrConnectionFailed
-	}
-
-	switch resp.StatusCode() {
-	case 404:
-		return nil, ErrNotFound
-	case 200:
-		return device, nil
-	default:
-		return nil, ErrUnknown
-	}
-}
-
-func buildURL(c *client, uri string) string {
-	u, _ := url.Parse(fmt.Sprintf("%s://%s:%d%s", c.scheme, c.host, c.port, uri))
-
-	return u.String()
 }
