+++ pkg/api/internalclient/session.go
@@ -1,0 +1,74 @@
+package internalclient
+
+import (
+	"fmt"
+
+	"github.com/shellhub-io/shellhub/pkg/models"
+)
+
+// sessionAPI defines methods for interacting with session-related functionality.
+type sessionAPI interface {
+	// SessionAsAuthenticated marks a session with the specified uid as authenticated.
+	// It returns a slice of errors encountered during the operation.
+	SessionAsAuthenticated(uid string) []error
+
+	// FinishSession finishes the session with the specified uid.
+	// It returns a slice of errors encountered during the operation.
+	FinishSession(uid string) []error
+
+	// KeepAliveSession sends a keep-alive signal for the session with the specified uid.
+	// It returns a slice of errors encountered during the operation.
+	KeepAliveSession(uid string) []error
+
+	// RecordSession records a session with the provided session information and record URL.
+	RecordSession(session *models.SessionRecorded, recordURL string)
+}
+
+func (c *client) SessionAsAuthenticated(uid string) []error {
+	var errors []error
+
+	_, err := c.http.
+		R().
+		SetBody(&models.Status{
+			Authenticated: true,
+		}).
+		Patch(fmt.Sprintf("/internal/sessions/%s", uid))
+	if err != nil {
+		errors = append(errors, err)
+	}
+
+	return errors
+}
+
+func (c *client) FinishSession(uid string) []error {
+	var errors []error
+
+	_, err := c.http.
+		R().
+		Post(fmt.Sprintf("/internal/sessions/%s/finish", uid))
+	if err != nil {
+		errors = append(errors, err)
+	}
+
+	return errors
+}
+
+func (c *client) KeepAliveSession(uid string) []error {
+	var errors []error
+
+	_, err := c.http.
+		R().
+		Post(fmt.Sprintf("/internal/sessions/%s/keepalive", uid))
+	if err != nil {
+		errors = append(errors, err)
+	}
+
+	return errors
+}
+
+func (c *client) RecordSession(session *models.SessionRecorded, recordURL string) {
+	_, _ = c.http.
+		R().
+		SetBody(session).
+		Post(fmt.Sprintf("http://"+recordURL+"/internal/sessions/%s/record", session.UID))
+}
